backend/requirements.txt
python-multipart
fastapi
python-docx
python-pptx
deep-translator
pymupdf
